Build list of report names based on file structure
There is no need to keep a manual list of report names in reports.json.
The previous implementation walked through a list of report names and
expected those to be folders under src/reports/. By using node's build
in methods to look for folder names, the list can be build
automatically.

This is implemented by using the buildin fs module. It lists all folders
in src/reports/ and uses these folder names as report names